Heart_Diseases_Prediction_LogisticRegression
A dataset is selected from the Kaggle repository with patient's medical history and attributes. By using this dataset, we predict whether the patient can have a heart disease or not. To predict this, we use  medical attributes of a patient and classify him if the patient is likely to have a heart disease.